+++ include/game/items.h
@@ -1,5 +1,5 @@
-#ifndef BULLET_H
-#define BULLET_H
+#ifndef ITEMS_H
+#define ITEMS_H
 
 #include "game.h"
 class Game;
@@ -7,16 +7,38 @@ class Game;
 class Item {
 private:
     Game* game;
+    H2DE_LevelPos pos;
+    H2DE_LevelVelocity velocity = { 0.0f, 0.0f };
 
+    std::string textureName;
+    int nbFrames;
     H2DE_LevelObject* object = nullptr;
+    bool pickedUp = false;
+    H2DE_Timeline* pickUpSpeedTimeline = nullptr;
+    float pickUpSpeed;
+    bool remove = false;
 
-    void initObject(std::string textureName);
+    void initVelocity();
+    void initObject();
+
+    void collided();
+    virtual void collidedImpl() = 0;
+    
+    virtual void updateImpl() = 0;
+    void updatePos();
+    void updateIndex();
+    void updateForCollisionWithPlayer();
 
 public:
-    Item(Game* game, H2DE_LevelPos pos, std::string textureName);
+    Item(Game* game, H2DE_LevelPos pos, std::string textureName, int nbFrames);
     virtual ~Item();
 
-    virtual void update() = 0;
+    void update();
+
+    void pickUp();
+
+    H2DE_LevelObjectData* getObjectData() const;
+    bool toRemove() const;
 };
 
 
@@ -26,11 +48,13 @@ private:
     Game* game;
     int level;
 
+    void collidedImpl() override;
+
+    void updateImpl() override;
+
 public:
     Xp(Game* game, H2DE_LevelPos pos, int level);
     ~Xp();
-
-    void update();
 };
 
 
@@ -39,11 +63,13 @@ class Coin : public Item {
 private:
     Game* game;
 
+    void collidedImpl() override;
+
+    void updateImpl() override;
+
 public:
     Coin(Game* game, H2DE_LevelPos pos);
     ~Coin();
-
-    void update();
 };
 
 #endif