+++ include/game/map.h
@@ -3,9 +3,11 @@
 
 #include "game.h"
 #include "entity.h"
+#include "items.h"
 class Game;
 class Player;
 class Enemy;
+class Item;
 class Bullet;
 
 class Map {
@@ -14,6 +16,7 @@ private:
     
     Player* player;
     std::vector<Enemy*> enemies;
+    std::vector<Item*> items;
     std::unordered_map<H2DE_LevelPos, Tile*> tiles;
 
     H2DE_Timeline* t = nullptr;
@@ -31,6 +34,7 @@ private:
 
     void updatePlayer();
     void updateEnemies();
+    void updateItems();
 
     void summonEnemy(int id, float size);
     void summonEnemy(int id, H2DE_LevelPos pos, float size);
@@ -43,10 +47,14 @@ public:
 
     void update();
 
+    void dropXp(H2DE_LevelPos pos, int level);
+    void dropCoin(H2DE_LevelPos pos);
+
     static int getIndex(float yPos, int index);
     Player* getPlayer() const;
     Enemy* getEnemy(H2DE_LevelObject* object) const;
     Bullet* getBullet(H2DE_LevelObject* object) const;
+    std::vector<Item*> getItems() const;
 };
 
 #endif