+++ include/game/utils.h
@@ -30,6 +30,7 @@ struct EntityData {
     EntityStats stats;
     std::optional<H2DE_LevelRect> collisionHitbox;
     H2DE_LevelRect damageHitbox;
+    int xpOnDeath;
 };
 
 struct TileGround {