+++ src/items/xp.cpp
@@ -1,7 +1,7 @@
 #include "items.h"
 
 // INIT
-Xp::Xp(Game* g, H2DE_LevelPos p, int l) : Item(g, p, "xp-lvl-" + std::to_string(l) + ".png"), level(l) {
+Xp::Xp(Game* g, H2DE_LevelPos p, int l) : Item(g, p, "xp-lvl-" + std::to_string(l) + ".png", 2), level(l) {
 
 }
 
@@ -10,7 +10,12 @@ Xp::~Xp() {
     std::cout << "Xp cleared" << std::endl;
 }
 
+// EVENTS
+void Xp::collidedImpl() {
+    
+}
+
 // UPDATE
-void Xp::update() {
+void Xp::updateImpl() {
 
 }